refactor to navbutton
