+++ components/Post/PostContainer.js
@@ -3,9 +3,9 @@ import {SafeAreaView, StyleSheet} from 'react-native';
 
 import {Post} from './Post';
 import {postService} from '../../services';
-import defaultStyles, {styleConstants} from '../../styles';
+import defaultStyles from '../../styles';
 import { defaultNavigationOptions } from '../../constants';
-import IconButton from '../common/IconButton';
+import NavButton from '../common/NavButton';
 
 export class PostContainer extends Component {
   constructor(props) {
@@ -20,19 +20,11 @@ export class PostContainer extends Component {
   static navigationOptions = props => ({
     ...defaultNavigationOptions,
     headerRight: (
-      <IconButton
-        onClick={() => {
-          props.navigation.navigate('AddComment', {
-            post: props.navigation.getParam('post')
-          })
-        }}
-        icon={{
-            name: 'plus',
-            size: 28,
-            color: styleConstants.palette.white
-        }}
-        containerStyle={styles.iconView}
-      />
+      <NavButton icon='plus' onClick={() => {
+        props.navigation.navigate('AddComment', {
+          post: props.navigation.getParam('post')
+        })
+      }}/>
     )
   })
 
@@ -82,8 +74,5 @@ export class PostContainer extends Component {
 const styles = StyleSheet.create({
   container: {
       padding: 10
-  },
-  iconView: {
-    marginRight: 10
   }
 });
