+++ components/PostList/CreatePostButton.js
@@ -1,22 +1,6 @@
 import React from 'react';
-import {StyleSheet} from 'react-native';
-import IconButton from '../common/IconButton';
-import { styleConstants } from '../../styles';
+import NavButton from '../common/NavButton';
 
 export default ({onClick}) => (
-    <IconButton
-        onClick={onClick}
-        icon={{
-            name: 'edit',
-            size: 28,
-            color: styleConstants.palette.white
-        }}
-        containerStyle={styles.iconView}
-    />
+    <NavButton onClick={onClick} icon='edit'/>
 )
-
-const styles = StyleSheet.create({
-    iconView: {
-        marginRight: 15
-    }
-});
